refactor: 파일명 컴포넌트명 불일치 수정
